Parser antigo movido para teste.rb
